Update index.html
Learned how to loop over an array to dynamically create list items

whenever working with form inputs we always want to bind them with v-model so that they change when clicked or altered

to avoid strange blips in the way your code works, always add a key to any element that has v-for on it

introduced to conditionals v-if and v-show. v-show toggles display

introduced to computed properties - sort of like methods that can be cached. We use these for less repetition in our code