Update index.html
If we want to create custom vue components, we can register them globally on the app or locally to parent vue components. Objects registered locally within vue components will take the exact same shape as parent vue components - they can have their own data methods, mounted hooks, ect.

Learned how to add templated css for vue components - adding slots if we want each instances text to be different or to reflect what is written in the ‘slot’ within the HTML. We can also use this template along with a data method to change the state of something, for example a button to disable it after it is clicked