index.html updated
Learned to take an objects presets that we want to have and put them into a JS export file, then import the object to our main file and reference them like we did before. However we cant import statements outside of a module, so we change our inline script tag to be type=“module” to alleviate that error.

One component per file and then import as needed for cleaner looking code

added vue chrome extension